requirements.txt
growwapi>=1.5.0
pyotp==2.9.0
supabase>=2.15.0
pandas==2.2.3
python-dotenv==1.0.1
